refactor to multiple plugins for onLoad and onResolve
