+++ src/index.tsx
@@ -1,8 +1,9 @@
 import * as esbuild from "esbuild-wasm";
 import { useState, useEffect, useRef } from "react";
-
 import { createRoot } from "react-dom/client";
+
 import { unpkgPathPlugin } from "./plugins/unpkg-path-plugin";
+import { fetchPlugin } from "./plugins/fetch-plugin";
 
 const App = () => {
   const [rawInput, setRawInput] = useState("");
@@ -28,19 +29,14 @@ const App = () => {
       return;
     }
 
-    // // transpile the rawInput to js code
-    // const result = await ref.current.transform(rawInput, {
-    //   loader: "jsx",
-    //   target: "es2015",
-    // });
-
     // transpile the rawInput to js code
     const result = await ref.current.build({
       // index.js will be the first file of bundling process
       entryPoints: ["index.js"],
       bundle: true,
       write: false,
-      plugins: [unpkgPathPlugin(rawInput)],
+      // plugins go from left to write onResolve -> onLoad
+      plugins: [unpkgPathPlugin(), fetchPlugin(rawInput)],
       define: {
         "process.env.NODE_ENV": "'production'",
         global: "window",
